@@ -11,8 +11,8 @@
     "ng": "ng",
     "start": "node server.js",
     "dev": "ng serve",
-    "build": "ng build --prod",
-    "gitbuild": "ng build --prod --base-href /dist/",
+    "build": "ng build --configuration production",
+    "gitbuild": "ng build --configuration production --base-href /dist/",
     "test": "ng test",
     "test-ci": "TEST_CI=true ng test",
     "lint": "eslint . --ext .ts --fix",
@@ -20,34 +20,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~0.1102.6",
-    "@angular/animations": "11.2.7",
-    "@angular/cli": "^11.2.6",
-    "@angular/common": "11.2.7",
-    "@angular/compiler": "11.2.7",
-    "@angular/compiler-cli": "^11.2.7",
-    "@angular/core": "11.2.7",
-    "@angular/forms": "11.2.7",
-    "@angular/localize": "^11.2.7",
-    "@angular/platform-browser": "11.2.7",
-    "@angular/platform-browser-dynamic": "11.2.7",
-    "@angular/router": "11.2.7",
-    "@ng-bootstrap/ng-bootstrap": "^9.1.0",
+    "@angular-devkit/build-angular": "^14.1.0",
+    "@angular/animations": "14.1.0",
+    "@angular/cli": "^14.1.0",
+    "@angular/common": "14.1.0",
+    "@angular/compiler": "14.1.0",
+    "@angular/compiler-cli": "^14.1.0",
+    "@angular/core": "14.1.0",
+    "@angular/forms": "14.1.0",
+    "@angular/localize": "^14.1.0",
+    "@angular/platform-browser": "14.1.0",
+    "@angular/platform-browser-dynamic": "14.1.0",
+    "@angular/router": "14.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^13.0.0",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^10.1.2",
-    "@ngrx/router-store": "^10.1.2",
-    "@ngrx/store": "^10.1.2",
-    "@ngrx/store-devtools": "^10.1.2",
+    "@ngrx/effects": "^14.0.2",
+    "@ngrx/router-store": "^14.0.2",
+    "@ngrx/store": "^14.0.2",
+    "@ngrx/store-devtools": "^14.0.2",
     "@ngx-translate/core": "13.0.0",
     "@ngx-translate/http-loader": "6.0.0",
-    "@okta/okta-angular": "^3.0.1",
+    "@okta/okta-angular": "^5.2.0",
+    "@okta/okta-auth-js": "^6.7.3",
+    "@popperjs/core": "^2.11.5",
     "angular-calendar": "^0.28.22",
     "chart.js": "2.9.4",
     "core-js": "3.8.2",
     "date-fns": "^2.16.1",
     "express": "^4.17.1",
     "font-awesome": "4.7.0",
-    "moment": "^2.29.1",
+    "moment": "^2.29.4",
     "ng2-charts": "^2.4.2",
     "ngx-currency": "^2.5.2",
     "ngx-ui-loader": "^10.0.0",
@@ -55,11 +57,11 @@
     "rxjs": "^6.6.6",
     "tslib": "^2.1.0",
     "yarn": "1.22.10",
-    "zone.js": "~0.11.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-eslint/eslint-plugin": "^1.1.0",
-    "@angular/language-service": "11.2.7",
+    "@angular/language-service": "14.1.0",
     "@types/jasmine": "3.6.2",
     "@types/jasminewd2": "2.0.8",
     "@types/node": "^14.14.20",
@@ -73,7 +75,7 @@
     "eslint-plugin-jsdoc": "^31.4.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~6.0.0",
-    "karma": "~5.2.3",
+    "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "2.0.0",
     "karma-coverage-istanbul-reporter": "~3.0.3",
@@ -82,7 +84,7 @@
     "protractor": "~7.0.0",
     "ts-node": "9.1.1",
     "tslint": "~6.1.0",
-    "typescript": "^4.1.3"
+    "typescript": "^4.7.4"
   },
   "heroku-run-build-script": true
 }
