@@ -10,10 +10,11 @@ import { ConfirmModule } from '../../shared/modules/confirm/confirm.module';
 import { NgbDateLocaleAdapter } from '../../shared/utils/ngb-date-locale.adapter';
 import { EditExpenseComponent } from './edit-expense/edit-expense.component';
 import { ExpenseComponent } from './expense.component';
+import { EditMultipleExpenseComponent } from './edit-multiple-expense/edit-multiple-expense.component'
 import { ExpenseRoutingModule } from './expense.routing.module';
 
 @NgModule({
-    declarations: [ExpenseComponent, EditExpenseComponent],
+    declarations: [ExpenseComponent, EditExpenseComponent, EditMultipleExpenseComponent],
     imports: [CommonModule,
         ExpenseRoutingModule,
         TranslateModule,
