@@ -1,13 +1,15 @@
 import {
  HttpErrorResponse, HttpEvent, HttpHandler, HttpHeaderResponse, HttpInterceptor, HttpRequest
 } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Inject, Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
+import { OKTA_AUTH } from '@okta/okta-angular';
+import { OktaAuth } from '@okta/okta-auth-js';
 import { NgxUiLoaderService } from 'ngx-ui-loader';
 import { Observable, throwError } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
-import { OktaAuthService } from '@okta/okta-angular';
+
 import { ShowAlertError } from '../../store/alert.actions';
 import { AppState } from '../../store/app.reducers';
 
@@ -16,12 +18,12 @@ export class HeaderInterceptor implements HttpInterceptor {
     constructor(private router: Router,
                 private store: Store<AppState>,
                 private ngxService: NgxUiLoaderService,
-                private oktaAuth: OktaAuthService) {
+                @Inject(OKTA_AUTH) private _oktaAuth: OktaAuth) {
     }
 
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         this.ngxService.start();
-        const token = this.oktaAuth.getAccessToken();
+        const token = this._oktaAuth.getAccessToken();
         const modified = req.clone({ setHeaders: { Authorization: `Bearer ${token}` } });
         return next.handle(modified).pipe(
             map((event: HttpHeaderResponse) => {
